Moved classes into their own files in appropriate packages.
