+++ src/gui_tools/DraggableNode.java
@@ -1,0 +1,139 @@
+package gui_tools;
+
+import javafx.event.EventHandler;
+import javafx.scene.Node;
+import javafx.scene.input.MouseEvent;
+import javafx.scene.layout.Pane;
+
+/**
+ * Simple draggable node.
+ * 
+ * Dragging code based on {@link http://blog.ngopal.com.np/2011/06/09/draggable-node-in-javafx-2-0/}
+ * 
+ * @author Michael Hoffer <info@michaelhoffer.de>
+ */
+public class DraggableNode extends Pane 
+{
+    // node position
+    private double x = 0;
+    private double y = 0;
+    // mouse position
+    private double mousex = 0;
+    private double mousey = 0;
+    private Node view;
+    private boolean dragging = false;
+    private boolean moveToFront = true;
+
+    public DraggableNode() 
+    {
+        init();
+    }
+
+    public DraggableNode(Node view) 
+    {
+        this.view = view;
+
+        getChildren().add(view);
+        init();
+    }
+
+    private void init() 
+    {
+        onMousePressedProperty().set(new EventHandler<MouseEvent>() 
+        {
+            @Override
+            public void handle(MouseEvent event) 
+            {
+                // record the current mouse X and Y position on Node
+                mousex = event.getSceneX();
+                mousey = event.getSceneY();
+
+                x = getLayoutX();
+                y = getLayoutY();
+
+                if (isMoveToFront()) 
+                {
+                    toFront();
+                }
+            }
+        });
+
+        //Event Listener for MouseDragged
+        onMouseDraggedProperty().set(new EventHandler<MouseEvent>() 
+        {
+            @Override
+            public void handle(MouseEvent event) 
+            {
+                // Get the exact moved X and Y
+                double offsetX = event.getSceneX() - mousex;
+                double offsetY = event.getSceneY() - mousey;
+
+                x += offsetX;
+                y += offsetY;
+
+                double scaledX = x;
+                double scaledY = y;
+
+                setLayoutX(scaledX);
+                setLayoutY(scaledY);
+
+                dragging = true;
+
+                // again set current Mouse x AND y position
+                mousex = event.getSceneX();
+                mousey = event.getSceneY();
+
+                event.consume();
+            }
+        });
+
+        onMouseClickedProperty().set(new EventHandler<MouseEvent>() 
+        {
+            @Override
+            public void handle(MouseEvent event) 
+            {
+
+                dragging = false;
+            }
+        });
+
+    }
+
+    /**
+     * @return the dragging
+     */
+    protected boolean isDragging() 
+    {
+        return dragging;
+    }
+
+
+    /**
+     * @return the view
+     */
+    public Node getView() 
+    {
+        return view;
+    }
+
+    /**
+     * @param moveToFront the moveToFront to set
+     */
+    public void setMoveToFront(boolean moveToFront) 
+    {
+        this.moveToFront = moveToFront;
+    }
+
+    /**
+     * @return the moveToFront
+     */
+    public boolean isMoveToFront() 
+    {
+        return moveToFront;
+    }
+    
+    public void removeNode(Node n) 
+    {
+        getChildren().remove(n);
+    }
+}
