+++ src/gui_tools/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Holds general classes and interfaces that extend the javFX library
+ */
+package gui_tools;
