+++ src/uml_class/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Holds classes relevant to the UML Class boxes that can be created and modified by the user.
+ */
+package uml_class;
